+++ aiofastnet/__init__.py
@@ -1,47 +1,44 @@
 import socket
 
-from .openssl_compat import OPENSSL_DYN_LIBS
-
+from .api_create_connection import create_connection
+from .api_create_datagram_endpoint import create_datagram_endpoint
+from .api_create_server import create_server
+from .api_create_unix_connection import create_unix_connection
+from .api_create_unix_server import create_unix_server
+from .api_patch import install_policy, loop_factory, patch_loop
+from .api_sendfile import sendfile
+from .api_start_tls import start_tls
 from .api_streams import (
     open_connection,
     start_server,
 )
-
-from .api_create_server import create_server
-from .api_create_connection import create_connection
-from .api_create_unix_connection import create_unix_connection
-from .api_create_unix_server import create_unix_server
-from .api_start_tls import start_tls
-from .api_sendfile import sendfile
-from .api_patch import loop_factory, patch_loop, install_policy
-
-from .transport import (
-    Transport,
-    Protocol,
-    aiofn_is_buffered_protocol
-)
+from .openssl_compat import OPENSSL_DYN_LIBS
+from .transport import Protocol, Transport, aiofn_is_buffered_protocol
 
 __all__ = [
     'OPENSSL_DYN_LIBS',
-    'open_connection',
-    'start_server',
-    'create_server',
+    'Protocol',
+    'Transport',
+    'aiofn_is_buffered_protocol',
     'create_connection',
+    'create_datagram_endpoint',
+    'create_server',
     'create_unix_connection',
     'create_unix_server',
-    'start_tls',
-    'sendfile',
-    'loop_factory',
-    'patch_loop',
     'install_policy',
-    'Transport',
-    'Protocol',
-    'aiofn_is_buffered_protocol',
+    'loop_factory',
+    'open_connection',
+    'patch_loop',
+    'sendfile',
+    'start_server',
+    'start_tls',
 ]
 
 if hasattr(socket, 'AF_UNIX'):
-    from .api_streams import (  # noqa: F401
+    from .api_streams import (
         open_unix_connection as open_unix_connection,
+    )
+    from .api_streams import (
         start_unix_server as start_unix_server,
     )
 
@@ -51,5 +48,5 @@ if hasattr(socket, 'AF_UNIX'):
     ))
 
 
-__version__ = "0.21.0"
+__version__ = "0.22.0"
 __author__ = "Taras Kozlov"