+++ aiofastnet/constants.py
@@ -18,15 +18,16 @@ SSL_BIO_SIZE_DEFAULTS = {
     # This is the size of the buffer that we pass to the recv syscall
     # The bigger it is the more we can read from kernel RCVBUF with a single syscall
     # But it also increases the memory footprint per client
-    "ssl_incoming_bio_size": int(16 * (16 * 1024 + 64)),
+    "ssl_incoming_bio_size": 16 * (16 * 1024 + 64),
 
     # Static size for the outgoing SSL BIO
     # Indicates how much encrypted data is accumulated before we call `send` syscall
     # Having extra 64 bytes prevents scenarios when we send almost complete TLS record.
     # It is not great for the latency.
-    "ssl_outgoing_bio_size": int(16 * (16 * 1024 + 64))
+    "ssl_outgoing_bio_size": 16 * (16 * 1024 + 64)
 }
 
 DATA_RECEIVED_MAX_SIZE = 256 * 1024
+DATAGRAM_RECEIVED_MAX_SIZE = 64 * 1024
 
 EXC_INFO_ATTR = '_aiofastnet_extra_info'